refactor: add side-filter, sort-bar and pagination constants
