+++ src/components/UI/molecules/Pagination/Pagination.stories.tsx
@@ -1,19 +1,11 @@
 import { Meta, StoryObj } from '@storybook/react'
 import { Pagination } from './Pagination.component'
+import { paginationProps } from '@constants'
 
 const meta: Meta<typeof Pagination> = {
   title: 'Molecules/ Pagination',
   component: Pagination,
-  args: {
-    total: 826,
-    current: 4,
-    pageSize: 20,
-    loading: false,
-    numberOfButtons: 5,
-    prevTitle: 'Anterior',
-    nextTitle: 'Siguiente',
-    onChange: (page: number, pageSize: number) => console.log('Pagination: ', { page, pageSize })
-  }
+  args: paginationProps
 }
 
 export default meta
