+++ src/components/UI/page/JobsPage/JobsPage.stories.tsx
@@ -2,26 +2,17 @@ import { Meta, StoryObj } from '@storybook/react'
 import React from 'react'
 import { MainButton } from '@components/UI/atoms'
 import JobsPage from './JobsPage.component'
-import {
-  IFilter,
-  IGetOptionsOnSearchProps,
-  ISearchRenderTypeOption,
-  ISetIsApplied,
-  ISideFilter,
-  IUnApplyWithChild
-} from '@components/UI/template'
-
+import { ISideFilter } from '@components/UI/template'
 import {
   CompanyLogo,
   MagnetoResolutionProps,
   RightsReservedProps,
-  appliedOptionsBySearchRenderType,
   listMenuText,
-  optionsFilterOrderBy,
   searchList,
   vacants
 } from '@constants/stories.constants'
-import { storiesFilters } from '@constants'
+import { sideFilterProps, sortBarProps, paginationProps } from '@constants'
+
 import { Buildings2, Export3 } from '@constants/icons.constants'
 
 const jobDetailsDrawer = {
@@ -182,45 +173,6 @@ const mobileJobDetailsDrawer = {
   }
 }
 
-const sortBarProps = {
-  textSortButton: 'FILTROS',
-  mainTitle: 'Ofertas de empleo en colombia',
-  filterSummary: '280 empleos',
-  setIsFiltersOpen: () => console.log('setIsFiltersOpen'),
-  orderFilter: () => console.log('orderFilter'),
-  orderFields: optionsFilterOrderBy,
-  textOrderFilter: 'Fecha de publicación',
-  setFilter: (menuItem: { key: string }) => console.log(menuItem.key),
-  orderByText: 'Organizar por',
-  isFiltersOpen: true,
-  loading: false
-}
-
-const sideFilterProps = {
-  title: 'Filtrar empleos',
-  filters: storiesFilters as IFilter[],
-  totalAppliedFilters: 1,
-  filterSummary: '3169 empleos',
-  buttonText: 'Limpiar',
-  loading: false,
-  isFiltersOpen: true,
-  switchText: 'Apto para personas con discapacidad',
-
-  setIsFiltersOpen: () => console.log('setIsFiltersOpen'),
-  setIsApplied: (filter: ISetIsApplied) => new Promise((resolve) => resolve(console.log({ filter }))),
-  clearFilters: () => new Promise((resolve) => resolve(console.log('clearFilters'))),
-  unApplyWithChild: (withChild: IUnApplyWithChild) => new Promise((resolve) => resolve(console.log({ withChild }))),
-  getOptionsOnLoad: (field: string, values: (string | number)[]) => {
-    console.log('getOptionsOnLoad: ', { field, values })
-    return new Promise<ISearchRenderTypeOption[]>((resolve) => resolve(appliedOptionsBySearchRenderType))
-  },
-  getOptionsOnSearch: (term: IGetOptionsOnSearchProps) =>
-    new Promise((resolve) => {
-      console.log('getOptionsOnSearch: ', { term })
-      return resolve([])
-    })
-}
-
 const frequentSearch = {
   searchHeading: 'Búsquedas de empleo más frecuentes',
   searchList,
@@ -233,17 +185,6 @@ const footerProps = {
   magnetoResolutionProps: MagnetoResolutionProps,
   rightsReservedProps: RightsReservedProps,
   menuFooterLink: listMenuText
-}
-
-const paginationProps = {
-  total: 826,
-  current: 4,
-  pageSize: 20,
-  loading: false,
-  numberOfButtons: 5,
-  prevTitle: 'Anterior',
-  nextTitle: 'Siguiente',
-  onChange: (page: number, pageSize: number) => console.log('Pagination: ', { page, pageSize })
 }
 
 const meta: Meta<typeof JobsPage> = {
