+++ src/components/UI/template/SortBar/SortBar.stories.tsx
@@ -1,25 +1,13 @@
 import React from 'react'
 import { Meta, StoryObj } from '@storybook/react'
 import SortBar from './SortBar.component'
-import { optionsFilterOrderBy } from '../../../../constants/stories.constants'
+import { sortBarProps } from '@constants'
 
 const meta: Meta<typeof SortBar> = {
   title: 'Template/Sort Bar',
   component: SortBar,
   tags: ['autodocs'],
-  args: {
-    textSortButton: 'FILTROS',
-    mainTitle: 'Ofertas de empleo en colombia',
-    filterSummary: '280 empleos',
-    orderFields: optionsFilterOrderBy,
-    textOrderFilter: 'Fecha de publicación',
-    orderByText: 'Organizar por',
-    isFiltersOpen: false,
-    loading: false,
-    setIsFiltersOpen: () => console.log('setIsFiltersOpen'),
-    orderFilter: () => console.log('orderFilter'),
-    setFilter: (menuItem: { key: string }) => console.log(menuItem.key)
-  },
+  args: sortBarProps,
   decorators: [
     (Story) => (
       <div style={{ margin: '-16px' }}>
