+++ src/constants/template/index.ts
@@ -1,1 +1,2 @@
 export * from './sideFilter.constants'
+export * from './sortBar.constants'
